+++ lib/src/proxy/editor.dart
@@ -216,7 +216,7 @@ class _EditorProxy extends _HasProxy implements Editor {
   
   void selectAll() => _proxy.selectAll();
   
-  String setValue(String value, {int cursorPosition: 0}) {
+  String setValue(String value, [int cursorPosition = 0]) {
     assert(cursorPosition >= -1 && cursorPosition <= 1);
     return _proxy.setValue(value, cursorPosition);
   }